+++ CurrencyConverterPro/Controllers/CurrencyConverterController.cs
@@ -39,7 +39,7 @@ namespace CurrencyConverterPro.Controllers
                     Amount = model.Amount
                 };
 
-                double convertedAmount = _logic.ConvertCurrency(currency);
+                decimal convertedAmount = _logic.ConvertCurrency(currency);
                 model.ConvertedResult = Math.Round(convertedAmount, 2);
 
                 return View("Index", model);
